CyclicArrays.CyclicArray — Type

CyclicArray

CyclicArray data structure. Available constructors:

CyclicArray(data::AbstractArray{T,N}, connections::AbstractArray)

To generate a new CyclicArray with the connections of x:

CyclicArray(x::CyclicArray)

To generate a new CyclicArray with the connections:

CyclicArray(connections)

To generate a new CyclicArray with the connections of y and data from x:

CyclicArray(x::AbstractArray,y::CyclicArray)

Base.diff — Function

diff

diff(A::CyclicArray)
diff(A::CyclicArray; dims::Integer)

Finite difference operator on a vector or a multidimensional array A. In the latter case the dimension to operate on needs to be specified with the dims keyword argument.

CyclicArrays.stagger — Function

stagger

stagger(A::CyclicArray; dims=1::Integer, frac=0.5::Real)

Linear interpolation of the array to intermidiate grid positions

CyclicArrays.shiftc — Function

shiftc shiftc(A::CyclicArray; dims=1::Integer, shift=1::Real) Shift array by integer